@@ -1,13 +1,21 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+//import ReactDOM from 'react-dom'; теперь react-dom/client в 18 версии
+import ReactDOM from 'react-dom/client';
 import App from './components/app/App';
-//import MarvelService from './services/MarvelService';
 import './style/style.scss';
 
-ReactDOM.render(
+ReactDOM
+    .createRoot(document.getElementById('root'))
+    .render(
+      <React.StrictMode>
+        <App />
+      </React.StrictMode>
+    )
+
+/* ReactDOM.render(
   <React.StrictMode>
     <App />
   </React.StrictMode>,
   document.getElementById('root')
-);
+); */
 
